+++ fabric/src/main/java/com/lx862/jcm/mixin/modded/mtr/VehicleResourceMixin.java
@@ -26,7 +26,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 @Mixin(value = VehicleResource.class, remap = false)
 public abstract class VehicleResourceMixin {
     @Inject(method = "queue(Lorg/mtr/mod/render/StoredMatrixTransformations;Lorg/mtr/mod/data/VehicleExtension;IIIZ)V", at = @At("HEAD"), cancellable = true)
-    private void jsblock$renderScriptResult(StoredMatrixTransformations storedMatrixTransformations, VehicleExtension vehicle, int carNumber, int totalCars, int light, boolean noOpenDoorways, CallbackInfo ci) {
+    private void jsblock$drawCarScript(StoredMatrixTransformations storedMatrixTransformations, VehicleExtension vehicle, int carNumber, int totalCars, int light, boolean noOpenDoorways, CallbackInfo ci) {
         // If hide riding vehicle and is current vehicle, both cancel the rendering, and cancel our script result rendering
         if(JCMClientConfig.INSTANCE.hideRidingVehicle.value() && VehicleRidingMovement.isRiding(vehicle.getId())) {
             ci.cancel();
@@ -40,8 +40,8 @@ public abstract class VehicleResourceMixin {
         ScriptInstance<?> scriptInstance = MTRContentScripting.getScriptManager().getInstanceManager().getInstance(new UniqueKey("vehicle", vehicle.getHexId(), scriptGroupId));
         if(!(scriptInstance instanceof VehicleScriptInstance)) return;
 
-        ScriptRenderManager carRenderManager = ((VehicleScriptInstance)scriptInstance).renderManagers.get(carNumber);
-        ScriptSoundManager carSoundManager = ((VehicleScriptInstance)scriptInstance).soundManagers.get(carNumber);
+        ScriptRenderManager carRenderManager = ((VehicleScriptInstance)scriptInstance).capturedScriptCalls.carRenderManagers.get(carNumber);
+        ScriptSoundManager carSoundManager = ((VehicleScriptInstance)scriptInstance).capturedScriptCalls.carSoundManagers.get(carNumber);
 
         StoredMatrixTransformations newTransform = storedMatrixTransformations.copy();
         newTransform.add(gh -> gh.translate(0, -1, 0)); // Replicate behaviour from MTR 3. Not sure if we should do it here tho?